Add a .gitignore to the proof directory
The common Makefile writes temporary files into the `gotos`, `logs`,
`html`, and `report` subdirectories of each proof directory by default.
This commit adds a .gitignore to be symlinked into the top-level proof
directory that ignores the contents of these directories, as
proof-writers will almost certainly want to have this file anyway.

In addition, the CBMC Viewer [1] tool, which is commonly used in conjunction
with this repository, emits temporary files that are prefixed with `TAGS-`.
This commit ignores those files too.

[1] https://github.com/awslabs/aws-viewer-for-cbmc