Do not remove function bodies during nondet-static
Prior to this commit, the common Makefile was removing function bodies
during the nondet-static pass. This is incorrect, because function
bodies should only be removed from the project object files, whereas the
nondet-static pass is applied to the linked binary (that contains the
proof stubs that stub out the removed functions). The remove-function
pass has already been applied earlier to the project binary, so it
should not be applied again to the linked binary.