Don't pass --unwindset when UNWINDSET not defined
This commit fixes the expansion of $CBMC_UNWINDSET. Previously, the
expansion of $CBMC_UNWINDSET had assumed that the $UNWINDSET variable
was defined but empty. However, this variable is not actually defined in
Makefile.common, not should it be.

This commit expands the --unwindset flag only if the $UNWINDSET variable
is both defined and empty.