Use a separate variable for proof ID
This commit introduces a new variable that proof-writers must fill in,
PROOF_ID. Previously, the unique identifier for the proof was equal to
the HARNESS_ENTRY variable, but this is inconsistent with how the
starter kit is meant to be used. After this commit, HARNESS_ENTRY shall
refer only to the name of the entry point of the harness, and will not
be used as a proof identifier.