Add generic "run all proofs in parallel" script
This commit adds a generic script that finds all proof directories and
adds a litani job for each of them, and then runs litani. This provides
a convenient way to run all CBMC proofs in parallel.

The script provides an option, `--no-standalone`, that allows the proofs
to be run as part of a larger set of litani jobs, for example running
all the proofs in multiple projects all at once.

This script should be usable as-is in many CBMC-using projects, without
further customization. Project owners need only add a value to the
PROJECT_NAME variable in Makefile-project-defines (this was already a
requirement even before this commit), so that the script can initialize
litani with the name of the project.