Allow proof to be run twice in multiple configs
This commit introduces a new variable that can be set in the environment
by the user, PROOF_SUFFIX. The value of variable gets appended to the
name of every proof. This allows the same proofs to be run multiple
times in different configurations in the same litani run. See the
documentation for an example.