Add Litani job to check proof coverage
Users can now define the EXPECTED_COVERAGE variable in proof Makefiles.
If this variable is defined, Litani will run the check-coverage program
on the value of EXPECTED_COVERAGE to determine whether the proof's
coverage was as high as expected.

If the EXPECTED_COVERAGE variable is not defined for a proof, the
coverage check is skipped. However, if the ENFORCE_COVERAGE_CHECKS
variable is defined for the whole repository, then proofs that have not
defined EXPECTED_COVERAGE will instead fail.