Make parallelism of EXPENSIVE proofs configurable
The run script now takes two flags to control whether and to what extent
to enable restricting the parallelism of expensive proofs.

--restrict-expensive-jobs controls whether to restrict the parallelism
of expensive jobs. By default it is enabled if Litani is a new enough
version to support it, and is disabled otherwise. Users can manually
specify "yes" or "no" to force a different behaviour.

--expensive-jobs-parallelism controls how many expensive jobs should be
run at once, if the feature is enabled.