Enable CBMC memory profile if Litani is new enough
The run script now checks whether Litani supports memory profiling
(which it does since version 1.8.0). If it does, and the user has not
disabled it by passing --no-memory-profile to the run script, the run
script will profile the CBMC safety check and coverage job. This means
that these two jobs will have a box-and-whisker diagram of memory usage
displayed on the front page, as well as a memory trace of the CBMC run
on the pipeline page.