Symlink rather than copy run-cbmc-proofs.py
Since this script appears to serve most users very well without
modification, we'll now symbolically link to it by default. Users who
need a customized run script can break the symlink manually and copy the
file over to modify it; in the future, we may also implement 'hooks' to
allow users to customize the script without copying it.