Change report generator to CBMC Viewer 2
All users are now expected to use version 2 of CBMC viewer, hosted at

  https://github.com/awslabs/aws-viewer-for-cbmc

This commit removes the rules for running version 1 of Viewer and
aliases their rules to the rules for running Viewer 2. This ensures that
neither the run-cbmc-proofs.py script, nor any other external scripts,
will break due to this change.