Add Litani job to check for function declarations
This commit adds a job that ensures that the link job does not complain
about missing function declarations. It checks for a certain warning in
the output of the link job; if this warning is printed, that indicates
that there is no declaration of the function under test in the proof.
This is a user error, so this commit ensures that the proof is aborted
if this happens.