Add a --target flag to run-cbmc-proofs.py
The value of this flag serves as the target that Make will be invoked on
when adding the Litani jobs for each proof. The old value, '_report', is
this flag's default value.

This is to prepare for a future commit where Makefile.common will invoke
the run script instead of manually running Litani.