Use run script even when running Make manually
As of this commit, running `make report` or another target in a proof
directory will run run-cbmc-proofs.py in the background, rather than
manually running Litani from the Makefile. The target to run is prefixed
with an underscore and passed to the run script, which will invoke make
in the proof directory only.

This ensures that the local proof-running experience matches CI as
closely as possible, since the run script contains sanity checks that
were not being run before this commit. As of this commit, developers who
run make in a proof directory (rather than running the run script) will
still get the benefits of the checks that the run script performs.