Add the repository module for the setup scripts.
The repository module provides methods that answer most of the
questions about the repository layout that the setup scripts have been
asking the user.  Examples are the location of the repository root,
the proof root, and the litani submodule.  In addition, the module can
use ctags to find the source files in the repository that define a
function with a given name.