Use the repository module in the setup scripts.
The repository module provides methods that answer most of the
questions about the repository layout that the setup scripts have been
asking the user.  In particular, we can find the set of source files
defining a given function and ask the user to choose the right file,
rather than asking the user to type it in by hand.