Add support for remove-static via crangler
crangler is a utility shipped with CBMC that permits (among other
features) selective removal of "static" from functions and objects via
JSON configuration files. From CBMC's perspective this is the preferred
route forward instead of using --export-file-local-symbols. Key
advantages are that selective removal is possible, and that there is no
renaming of functions. The latter makes for more intuitive proof
harnesses.