Invoke viewer with absolute path to cbmc-viewer.json
This patch adds to the viewer invocation the --config option with the
absolute path to the cbmc-viewer.json configuration file.

The run-cbmc-proofs script invokes viewer from the top-level proof
root with the absolute paths to the cbmc output.  Viewer by default
looks for the cbmc-viewer.json configuration file in the current
working directory.  This default behavior works when viewer is invoked
in the proof directory, but not in the proof root.  This patch fixes
this problem by overriding the default behavior and giving the
absolute path to the file.