Rewrite setup scripts to copy and not link.
This is a nearly total rewrite of setup.py, setup_proof.py and
util.py.  First, they are written to copy template files into place
instead of symlinking to them in the submodule.  Second, they are
rewritten to use the GitPython and Path modules.  Both changes
resulted in much less, much simpler code.