Remove support for make-source
The make-source tool is a command-line interface to the methods that
viewer can use to generate the list of source files it scans for
symbols with ctags.  Viewer links these symbols in a trace to their
definitions in the source.

By default, make-source scans the symbol table of the goto binary for
the source files defining symbols in the table.  Alternatively,
make-source (like Meta Infer) can use the process of building the goto
binary to find files like header files used in the build that do not
appear in the symbol table.  But this method can be brittle.

This commit removes support for scanning the build process for source
files, leaving only the default support for scanning the symbol table.