Add support for reports without coverage
The "make report" target does both property checking and coverage
checking and produces a report with both.  The report appears in the
subdirectory "report". But coverage checking can be slow.  This commit
adds a "make report-no-coverage" target that omits coverage checking,
and produces a report with only the results of property checking.  The
report appears in the subdirectory "report-no-coverage".