Add a COVERAGE variable for the report-no-coverage target
The report-no-coverage target was omitting coverage from the report,
but was still doing coverage checking and building coverage.xml.  The
problem was that the report-no-coverage target called the report
target, and report had coverage.xml hard-coded as a dependency.  This
patch replaces that hard-coded dependency with $(COVERAGE), and ensures
COVERAGE="" for report-no-coverage and COVERAGE="coverage.xml" for
report.