Update "summarize" utility and allow it to be called as a script
The script offers the option to add a column to the tabular summaries
pointing to the CBMC viewer HTML reports, if those have previously
been uploaded to a S3 bucket that is associated to a CloudFront
distribution. This use case serves a purpose for a GitHub Actions
workflow to be introduced in a forthcoming commit. The script also
offers to print the tabular summaries, just like before, if no AWS
infrastructure has been provisioned.